Allow checking out remote branches from the commit list
This commit allows to check out remote branches from the context menu
of a commit.

When checking out a remote branch and a local branch of the same name
already exists, a dialog is shown that offers the choice between
resetting the existing branch to that commit and checking out a detached
head.

From André Hänsel for PR #167