Fix performance regression on viewing status diffs with large files
The regression was introduced by getting the index from the repo instead of
passing it around along with the status diff. This caused the staged cache
to be discarded immediately after each query. The fix is to tie the lifetime
of the index and its cache directly to the lifetime of the status diff.