Fix crash on LFS-not-found error emitted from a different thread
This can happen, for example, from the config dialog when opening a
repo which has had LFS intiailized on it, but git-lfs can't be found.

The crash happens because the log entry can't be created with a parent
on a different thread, so it gets created unparented. The fix is to use
the connect overload that takes a context so that the connection type
can be automatic and the connection can be queued when the current
thread differs from the context's thread.

Fixes #247